Move function to better location
